tech: renommer deleteAttachment en archiveAttachment côté service
