+++ packages/api/server/models/attachmentModel/index.ts
@@ -1,11 +1,9 @@
 import createLinkedAttachment from './createLinkedAttachment';
-import deleteAttachment from './deleteAttachment';
 import findKeys from './findKeys';
 import serializeAttachment from './serializeAttachment';
 
 export default {
     createLinkedAttachment,
-    deleteAttachment,
     findKeys,
     serializeAttachment,
 };
