+++ packages/api/server/services/attachment/archiveAttachment.ts
@@ -1,10 +1,10 @@
 import { AttachmentKeys } from '#server/models/attachmentModel/findKeys';
-import attachmentModel from '#server/models/attachmentModel';
 import ServiceError from '#server/errors/ServiceError';
+import archiveAttachment from '#server/models/attachmentModel/archiveAttachment';
 
 export default async (keys: AttachmentKeys): Promise<void> => {
     try {
-        await attachmentModel.deleteAttachment(keys.attachment_id);
+        await archiveAttachment(keys.attachment_id);
     } catch (e) {
         throw new ServiceError('delete_failed', e);
     }
