+++ packages/api/server/services/attachment/index.ts
@@ -1,4 +1,4 @@
-import deleteAttachment from './deleteAttachment';
+import deleteAttachment from './archiveAttachment';
 import upload from './upload';
 
 export default {
